+++ src/main/java/org/loose/fis/sre/controllers/BookingController.java
@@ -45,7 +45,7 @@ public class BookingController {
         day.getItems().addAll("1","2","3","4","5","6","7","8","9","10","11","12","13","14","15","16","17","18",
                 "19","20","21","22","23","24","25","26","27","28","29","30","31");
         month.getItems().addAll("January","February","March","April","May","June","July","August","September","October","November","December");
-        year.getItems().addAll("2021","2022");
+        year.getItems().addAll("2021","2022","2023","2024");
         hour.getItems().addAll("9:00", "10:00", "11:00", "12:00", "13:00", "14:00", "15:00", "16:00");
         agent_book.getItems().addAll("Angela", "Agnes" ,"Angelo", "Agel");
     }
