+++ src/main/java/org/loose/fis/sre/controllers/BuyerManuController.java
@@ -17,21 +17,7 @@ import java.io.IOException;
 public class BuyerManuController {
 
 
-    @FXML
-    public void handlebook() {
 
-        try {
-            Parent root;
-            root = FXMLLoader.load(getClass().getClassLoader().getResource("book.fxml"));
-            Stage stage = new Stage();
-            stage.setTitle("Agentie Imobiliara");
-            stage.setScene(new Scene(root, 600, 575));
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-    }
     @FXML
     public void SeeHouses() {
