+++ src/main/resources/buyer_menu.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.text.*?>
 
 <GridPane alignment="center" hgap="11" prefHeight="500" prefWidth="700.0" vgap="11" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1"
-          >
+          fx:controller="org.loose.fis.sre.controllers.BuyerManuController" >
     <padding>
         <Insets bottom="25.0" left="25.0" right="25.0" top="25.0" />
     </padding>
