+++ src/main/resources/see_houses.fxml
@@ -1,26 +1,48 @@
+<?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.geometry.*?>
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
 
-<GridPane xmlns:fx="http://javafx.com/fxml/1" alignment="center" hgap="11" prefHeight="500" prefWidth="700"
-          vgap="11" xmlns="http://javafx.com/javafx/10.0.2-internal"
+<GridPane alignment="center" hgap="11" prefHeight="500" prefWidth="700" vgap="11" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1"
           fx:controller="org.loose.fis.sre.controllers.SeeHousesController">
     <padding>
         <Insets bottom="25.0" left="25.0" right="25.0" top="25.0" />
     </padding>
     <Text id="title-text" text="Agentie Imobiliara" wrappingWidth="124.125" GridPane.columnIndex="0" GridPane.columnSpan="2" GridPane.rowIndex="0" />
 
-    <Text fx:id="seehousemesage" GridPane.columnIndex="1" GridPane.rowIndex="10"/>
-    <HBox alignment="bottom_right" spacing="10" GridPane.columnIndex="1" GridPane.rowIndex="3">
-        <Button  onAction="#handleSeeHouses" text="See all the houses on the market"/>
-    </HBox>
-    <HBox alignment="bottom_right" spacing="10" GridPane.columnIndex="2" GridPane.rowIndex="3">
-        <Button   onAction="#handleSearchHouseAction" text="Search a house"/>
-    </HBox>
-    <TextField fx:id="Address" GridPane.columnIndex="3" GridPane.rowIndex="3"/>
 
-    <Label text="Address:" GridPane.columnIndex="3" GridPane.rowIndex="2"/>
+    <HBox alignment="bottom_right" spacing="10" GridPane.columnIndex="0" GridPane.rowIndex="3">
+        <Button onAction="#handleSeeHouses" prefHeight="26.0" prefWidth="188.0" text="See houses on the market" />
+    </HBox>
+    <HBox alignment="bottom_right" spacing="10" GridPane.columnIndex="0" GridPane.rowIndex="4">
+        <Button onAction="#handleSearchHouseAction" prefHeight="26.0" prefWidth="188.0" text="Search a house by address" />
+    </HBox>
+    <HBox alignment="bottom_right" spacing="10" GridPane.columnIndex="0" GridPane.rowIndex="5">
+        <Button onAction="#handlebook" prefHeight="26.0" prefWidth="188.0" text="Book a visit for a house" />
+    </HBox>
+    <TextField fx:id="Address" GridPane.columnIndex="2" GridPane.rowIndex="4" />
+
+    <Label text="Address:" GridPane.columnIndex="1" GridPane.rowIndex="4" />
+    <Text fx:id="seehousemesage" GridPane.columnIndex="1" GridPane.rowIndex="10" />
+   <columnConstraints>
+      <ColumnConstraints />
+      <ColumnConstraints />
+      <ColumnConstraints />
+   </columnConstraints>
+   <rowConstraints>
+      <RowConstraints />
+      <RowConstraints />
+      <RowConstraints />
+      <RowConstraints />
+      <RowConstraints />
+      <RowConstraints />
+      <RowConstraints />
+      <RowConstraints />
+      <RowConstraints />
+      <RowConstraints />
+      <RowConstraints />
+   </rowConstraints>
 
 </GridPane>
